// web/src/api-access/generated/models/userRoleResponseDto.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Unified.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface UserRoleResponseDto {
  id?: number;
  userId?: string;
  roleId?: number;
  effectiveDate?: string;
  /** @nullable */
  expiryDate?: string | null;
  /** @nullable */
  expiryReason?: string | null;
}